+++ CHANGELOG.md
@@ -13,9 +13,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Add head_js block into base html template
-- list teacher's organizations in the teacher dashboard sidebar.
+- List teacher's courses in the teacher courses dashboard page.
 - Added Certificates in the dashboard
 - Dashboard infinite scroll of orders and enrollments
+- list teacher's courses in the teacher courses dashboard page.
 
 ### Fixed